Extract glow lichen multi face logic into traits (#6396)
This will be useful for future block additions
